+++ componentapiexample/src/main/java/net/gini/android/vision/component/MainActivity.java
@@ -2,19 +2,22 @@ package net.gini.android.vision.component;
 
 import static net.gini.android.vision.example.ExampleUtil.isIntentActionViewOrSend;
 
+import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.Button;
+import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import net.gini.android.GiniApiType;
 import net.gini.android.vision.DocumentImportEnabledFileTypes;
 import net.gini.android.vision.GiniVision;
 import net.gini.android.vision.GiniVisionDebug;
@@ -31,6 +34,9 @@ import net.gini.android.vision.requirements.RequirementsReport;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Entry point for the component api example app.
+ */
 public class MainActivity extends AppCompatActivity {
 
     private Button mButtonStartGiniVisionCompat;
@@ -39,6 +45,8 @@ public class MainActivity extends AppCompatActivity {
     private RuntimePermissionHandler mRuntimePermissionHandler;
     private TextView mTextAppVersion;
     private TextView mTextGiniVisionLibVersion;
+    private Spinner mGiniApiTypeSpinner;
+    private GiniApiType mGiniApiType = GiniApiType.DEFAULT;
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
@@ -73,18 +81,22 @@ public class MainActivity extends AppCompatActivity {
 
     private void initGiniVision() {
         final BaseExampleApp app = (BaseExampleApp) getApplication();
-        // Configure the Gini Vision Library
         GiniVision.cleanup(this);
-        GiniVision.newInstance()
-                .setGiniVisionNetworkService(app.getGiniVisionNetworkService("ComponentAPI"))
-                .setGiniVisionNetworkApi(app.getGiniVisionNetworkApi())
-                .setDocumentImportEnabledFileTypes(DocumentImportEnabledFileTypes.PDF_AND_IMAGES)
-                .setFileImportEnabled(true)
-                .setQRCodeScanningEnabled(true)
-                .setMultiPageEnabled(true)
-                // Uncomment to add an extra page to the Onboarding pages
-//                .setCustomOnboardingPages(getOnboardingPages())
-                .build();
+        app.clearGiniVisionNetworkInstances();
+        final GiniVision.Builder builder = GiniVision.newInstance()
+                .setGiniVisionNetworkService(
+                        app.getGiniVisionNetworkService("ComponentAPI",
+                                mGiniApiType)
+                ).setGiniVisionNetworkApi(app.getGiniVisionNetworkApi());
+        if (mGiniApiType == GiniApiType.DEFAULT) {
+            builder.setDocumentImportEnabledFileTypes(DocumentImportEnabledFileTypes.PDF_AND_IMAGES)
+                    .setFileImportEnabled(true)
+                    .setQRCodeScanningEnabled(true)
+                    .setMultiPageEnabled(true);
+        }
+        // Uncomment to add an extra page to the Onboarding pages
+//      builder.setCustomOnboardingPages(getOnboardingPages());
+        builder.build();
     }
 
     private ArrayList<OnboardingPage> getOnboardingPages() {
@@ -114,11 +126,6 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void startGiniVisionStandardForImportedFile(@NonNull final Intent importedFileIntent) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.JELLY_BEAN_MR1) {
-            Toast.makeText(this, R.string.component_api_min_sdk_level_error,
-                    Toast.LENGTH_SHORT).show();
-            return;
-        }
         mRuntimePermissionHandler.requestStoragePermission(
                 new RuntimePermissionHandler.Listener() {
                     @Override
@@ -167,6 +174,18 @@ public class MainActivity extends AppCompatActivity {
                 startGiniVisionCompat();
             }
         });
+        mGiniApiTypeSpinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(final AdapterView<?> parent, final View view,
+                    final int position, final long id) {
+                mGiniApiType = GiniApiType.values()[position];
+            }
+
+            @Override
+            public void onNothingSelected(final AdapterView<?> parent) {
+
+            }
+        });
     }
 
     private void startGiniVisionCompat() {
@@ -196,11 +215,6 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void startGiniVisionStandard() {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.JELLY_BEAN_MR1) {
-            Toast.makeText(this, R.string.component_api_min_sdk_level_error,
-                    Toast.LENGTH_SHORT).show();
-            return;
-        }
         initGiniVision();
         mRuntimePermissionHandler.requestCameraPermission(
                 new RuntimePermissionHandler.Listener() {
@@ -252,6 +266,7 @@ public class MainActivity extends AppCompatActivity {
         mButtonStartGiniVisionCompat = (Button) findViewById(R.id.button_start_gini_vision_compat);
         mTextGiniVisionLibVersion = (TextView) findViewById(R.id.text_gini_vision_version);
         mTextAppVersion = (TextView) findViewById(R.id.text_app_version);
+        mGiniApiTypeSpinner = findViewById(R.id.gini_api_type_spinner);
     }
 
     private void createRuntimePermissionsHandler() {
@@ -274,6 +289,7 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
+    @SuppressLint("SetTextI18n")
     private void showVersions() {
         mTextGiniVisionLibVersion.setText(
                 "Gini Vision Library v" + net.gini.android.vision.BuildConfig.VERSION_NAME);