+++ ginivision-accounting-network/src/main/java/net/gini/android/vision/accounting/network/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Implementations of the {@link net.gini.android.vision.network.GiniVisionNetworkService}
+ * and the {@link net.gini.android.vision.network.GiniVisionNetworkApi} using the Gini Accounting API.
+ */
+package net.gini.android.vision.accounting.network;