+++ screenapiexample/src/main/java/net/gini/android/vision/screen/MainActivity.java
@@ -2,6 +2,7 @@ package net.gini.android.vision.screen;
 
 import static net.gini.android.vision.example.ExampleUtil.isPay5Extraction;
 
+import android.annotation.SuppressLint;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,10 +10,13 @@ import android.support.annotation.NonNull;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.Button;
+import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import net.gini.android.GiniApiType;
 import net.gini.android.vision.AsyncCallback;
 import net.gini.android.vision.DocumentImportEnabledFileTypes;
 import net.gini.android.vision.GiniVision;
@@ -39,6 +43,9 @@ import ch.qos.logback.classic.LoggerContext;
 import ch.qos.logback.classic.android.LogcatAppender;
 import ch.qos.logback.classic.encoder.PatternLayoutEncoder;
 
+/**
+ * Entry point for the screen api example app.
+ */
 public class MainActivity extends AppCompatActivity {
 
     public static final String EXTRA_OUT_EXTRACTIONS = "EXTRA_OUT_EXTRACTIONS";
@@ -51,7 +58,9 @@ public class MainActivity extends AppCompatActivity {
     private RuntimePermissionHandler mRuntimePermissionHandler;
     private TextView mTextGiniVisionLibVersion;
     private TextView mTextAppVersion;
+    private Spinner mGiniApiTypeSpinner;
     private CancellationToken mFileImportCancellationToken;
+    private GiniApiType mGiniApiType = GiniApiType.DEFAULT;
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
@@ -187,6 +196,7 @@ public class MainActivity extends AppCompatActivity {
                 || Intent.ACTION_SEND_MULTIPLE.equals(action);
     }
 
+    @SuppressLint("SetTextI18n")
     private void showVersions() {
         mTextGiniVisionLibVersion.setText(
                 "Gini Vision Library v" + net.gini.android.vision.BuildConfig.VERSION_NAME);
@@ -205,6 +215,18 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(final View v) {
                 startGiniVisionLibrary();
+            }
+        });
+        mGiniApiTypeSpinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(final AdapterView<?> parent, final View view,
+                    final int position, final long id) {
+                mGiniApiType = GiniApiType.values()[position];
+            }
+
+            @Override
+            public void onNothingSelected(final AdapterView<?> parent) {
+
             }
         });
     }
@@ -283,22 +305,27 @@ public class MainActivity extends AppCompatActivity {
     private void configureGiniVision() {
         final BaseExampleApp app = (BaseExampleApp) getApplication();
         GiniVision.cleanup(this);
-        GiniVision.newInstance()
-                .setGiniVisionNetworkService(app.getGiniVisionNetworkService("ScreenAPI"))
-                .setGiniVisionNetworkApi(app.getGiniVisionNetworkApi())
-                .setDocumentImportEnabledFileTypes(DocumentImportEnabledFileTypes.PDF_AND_IMAGES)
-                .setFileImportEnabled(true)
-                .setQRCodeScanningEnabled(true)
-                .setMultiPageEnabled(true)
-                // Uncomment to add an extra page to the Onboarding pages
-//                .setCustomOnboardingPages(getOnboardingPages())
-                // Uncomment to disable automatically showing the OnboardingActivity the
-                // first time the CameraActivity is launched - we highly recommend letting the
-                // Gini Vision Library show the OnboardingActivity at first run
-//                .setShouldShowOnboardingAtFirstRun(false)
-                // Uncomment to show the OnboardingActivity every time the CameraActivity starts
-//                .setShouldShowOnboarding(true)
-                .build();
+        app.clearGiniVisionNetworkInstances();
+        final GiniVision.Builder builder = GiniVision.newInstance()
+                .setGiniVisionNetworkService(
+                        app.getGiniVisionNetworkService("ScreenAPI",
+                                mGiniApiType)
+                ).setGiniVisionNetworkApi(app.getGiniVisionNetworkApi());
+        if (mGiniApiType == GiniApiType.DEFAULT) {
+            builder.setDocumentImportEnabledFileTypes(DocumentImportEnabledFileTypes.PDF_AND_IMAGES)
+                    .setFileImportEnabled(true)
+                    .setQRCodeScanningEnabled(true)
+                    .setMultiPageEnabled(true);
+        }
+        // Uncomment to add an extra page to the Onboarding pages
+//                builder.setCustomOnboardingPages(getOnboardingPages());
+        // Uncomment to disable automatically showing the OnboardingActivity the
+        // first time the CameraActivity is launched - we highly recommend letting the
+        // Gini Vision Library show the OnboardingActivity at first run
+//                builder.setShouldShowOnboardingAtFirstRun(false);
+        // Uncomment to show the OnboardingActivity every time the CameraActivity starts
+//                builder.setShouldShowOnboarding(true);
+        builder.build();
     }
 
     private void showUnfulfilledRequirementsToast(final RequirementsReport report) {
@@ -325,6 +352,7 @@ public class MainActivity extends AppCompatActivity {
         mButtonStartScanner = (Button) findViewById(R.id.button_start_scanner);
         mTextGiniVisionLibVersion = (TextView) findViewById(R.id.text_gini_vision_version);
         mTextAppVersion = (TextView) findViewById(R.id.text_app_version);
+        mGiniApiTypeSpinner = findViewById(R.id.gini_api_type_spinner);
     }
 
     private ArrayList<OnboardingPage> getOnboardingPages() {
@@ -380,9 +408,9 @@ public class MainActivity extends AppCompatActivity {
                     final GiniVisionError error = data.getParcelableExtra(
                             CameraActivity.EXTRA_OUT_ERROR);
                     if (error != null) {
-                        Toast.makeText(this, "Error: " +
-                                        error.getErrorCode() + " - " +
-                                        error.getMessage(),
+                        Toast.makeText(this, "Error: "
+                                        + error.getErrorCode() + " - "
+                                        + error.getMessage(),
                                 Toast.LENGTH_LONG).show();
                     }
                     break;